Move ObjectQuery to object/ dir
