+++ packages/client/src/observable/internal/BaseCollectionQuery.ts
@@ -17,7 +17,7 @@
 import type { InterfaceHolder } from "../../object/convertWireToOsdkObjects/InterfaceHolder.js";
 import type { ObjectHolder } from "../../object/convertWireToOsdkObjects/ObjectHolder.js";
 import type { Status } from "../ObservableClient/common.js";
-import { type ObjectCacheKey } from "./ObjectCacheKey.js";
+import { type ObjectCacheKey } from "./object/ObjectCacheKey.js";
 
 /**
  * Abstract base for ListQuery and SpecificLinkQuery.
