+++ packages/client/src/observable/internal/CacheKeys.ts
@@ -21,7 +21,7 @@ import type { CacheKey } from "./CacheKey.js";
 import type { KnownCacheKey } from "./KnownCacheKey.js";
 import type { SpecificLinkCacheKey } from "./links/SpecificLinkCacheKey.js";
 import type { ListCacheKey } from "./list/ListCacheKey.js";
-import type { ObjectCacheKey } from "./ObjectCacheKey.js";
+import type { ObjectCacheKey } from "./object/ObjectCacheKey.js";
 import type { OrderByCanonicalizer } from "./OrderByCanonicalizer.js";
 import type { WhereClauseCanonicalizer } from "./WhereClauseCanonicalizer.js";
