+++ packages/client/src/observable/internal/Changes.ts
@@ -20,7 +20,7 @@ import type { CacheKey } from "./CacheKey.js";
 import { DEBUG_ONLY__cacheKeyToString } from "./CacheKey.js";
 import type { SpecificLinkCacheKey } from "./links/SpecificLinkCacheKey.js";
 import type { ListCacheKey } from "./list/ListCacheKey.js";
-import type { ObjectCacheKey } from "./ObjectCacheKey.js";
+import type { ObjectCacheKey } from "./object/ObjectCacheKey.js";
 
 export class Changes {
   modifiedObjects: MultiMap<string, ObjectHolder> = new MultiMap();
