+++ packages/client/src/observable/internal/KnownCacheKey.ts
@@ -16,7 +16,7 @@
 
 import type { SpecificLinkCacheKey } from "./links/SpecificLinkCacheKey.js";
 import type { ListCacheKey } from "./list/ListCacheKey.js";
-import type { ObjectCacheKey } from "./ObjectCacheKey.js";
+import type { ObjectCacheKey } from "./object/ObjectCacheKey.js";
 
 export type KnownCacheKey =
   | ObjectCacheKey
