+++ packages/client/src/observable/internal/Store.test.ts
@@ -54,7 +54,7 @@ import type {
   ObserveListOptions,
   Unsubscribable,
 } from "../ObservableClient.js";
-import type { ObjectCacheKey } from "./ObjectCacheKey.js";
+import type { ObjectCacheKey } from "./object/ObjectCacheKey.js";
 import { createOptimisticId } from "./OptimisticId.js";
 import { runOptimisticJob } from "./OptimisticJob.js";
 import { Store } from "./Store.js";
