+++ packages/client/src/observable/internal/createCollectionConnectable.test.ts
@@ -22,7 +22,7 @@ import type { ObjectHolder } from "../../object/convertWireToOsdkObjects/ObjectH
 import type { Status } from "../ObservableClient/common.js";
 import type { CacheKey } from "./CacheKey.js";
 import { createCollectionConnectable as actualCreateCollectionConnectable } from "./createCollectionConnectable.js";
-import type { ObjectCacheKey } from "./ObjectCacheKey.js";
+import type { ObjectCacheKey } from "./object/ObjectCacheKey.js";
 import type { Store, SubjectPayload } from "./Store.js";
 import {
   createClientMockHelper,
