+++ packages/client/src/observable/internal/list/ListQuery.ts
@@ -43,7 +43,7 @@ import type { Canonical } from "../Canonical.js";
 import { type Changes, DEBUG_ONLY__changesToString } from "../Changes.js";
 import type { Entry } from "../Layer.js";
 import type { ListQueryOptions } from "../ListQueryOptions.js";
-import { type ObjectCacheKey } from "../ObjectCacheKey.js";
+import { type ObjectCacheKey } from "../object/ObjectCacheKey.js";
 import { objectSortaMatchesWhereClause as objectMatchesWhereClause } from "../objectMatchesWhereClause.js";
 import type { OptimisticId } from "../OptimisticId.js";
 import type { SimpleWhereClause } from "../SimpleWhereClause.js";
