+++ packages/client/src/observable/internal/object/ObjectCacheKey.ts
@@ -15,9 +15,9 @@
  */
 
 import type { ObjectTypeDefinition, PrimaryKeyType } from "@osdk/api";
-import type { ObjectHolder } from "../../object/convertWireToOsdkObjects/ObjectHolder.js";
-import type { CacheKey } from "./CacheKey.js";
-import type { ObjectQuery } from "./object/ObjectQuery.js";
+import type { ObjectHolder } from "../../../object/convertWireToOsdkObjects/ObjectHolder.js";
+import type { CacheKey } from "../CacheKey.js";
+import type { ObjectQuery } from "./ObjectQuery.js";
 
 export interface ObjectCacheKey extends
   CacheKey<
